ADDED: write to dtf method
The only issue I could not decide on was whether to end each write to do dtf with an empty transaction, end of dtf and then remove it every time the next user writes to the dtf. Or only use it once a user calls 'quit' without being logged in thus terminating the program.